pyproject.toml
[project]
name = "whisper-wire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.116.1",
    "passlib>=1.7.4",
    "pendulum>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "typer>=0.17.4",
]
